requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
altgraph==0.17.4
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
attrs==24.3.0
Babel==2.8.1
backports.tarfile==1.1.1
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.8.2
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
comtypes==1.4.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
docutils==0.21.2
Eel==0.16.0
evdev==1.7.1
exceptiongroup==1.2.1
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.16.1
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
fonttools==4.55.3
fqdn==1.5.1
fsspec==2024.12.0
future==1.0.0
gast==0.5.4
gevent==24.2.1
gevent-websocket==0.10.1
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.63.0
gruut==1.3.1
gruut-ipa==0.10.1
gruut_lang_de==1.3.1
gruut_lang_es==1.3.1
gruut_lang_fr==1.3.1
gruut_lang_it==1.3.1
gruut_lang_nl==1.3.1
gruut_lang_ru==1.3.1
gruut_lang_sv==1.3.1
gruut_lang_sw==1.3.1
gTTS==2.5.1
h11==0.14.0
h2==4.1.0
h5py==3.11.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
humanfriendly==10.0
Hypercorn==0.16.0
hyperframe==6.0.1
idna==3.7
ipykernel==6.29.5
ipython==8.31.0
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.context==5.3.0
jedi==0.19.2
Jinja2==3.1.4
json5==0.10.0
jsonlines==1.2.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.3.3
keyboard==0.13.5
Kivy==2.2.1
Kivy-Garden==0.1.5
kiwisolver==1.4.8
larynx==0.5.0
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.0
ml-dtypes==0.3.2
more-itertools==10.2.0
MouseInfo==0.1.3
mpmath==1.3.0
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook_shim==0.2.4
num2words==0.5.13
numpy==1.26.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnxruntime==1.17.3
opt-einsum==3.3.0
optree==0.11.0
overrides==7.7.0
packaging==24.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.3.6
playsound==1.3.0
priority==2.0.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
protobuf==4.25.3
psutil==5.8.0
ptyprocess==0.7.0
pure_eval==0.2.3
pvporcupine==3.0.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycairo==1.26.0
pycparser==2.22
pycryptodome==3.21.0
pydbus==0.6.0
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.18.0
PyGObject==3.48.2
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
PyMsgBox==1.0.9
pynput==1.7.7
pyparsing==3.1.2
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-json-logger==3.2.1
python-pidfile==3.0.0
python-version==0.0.2
python-xlib==0.33
python3-xlib==0.15
pyttsx==1.1
pyttsx3==2.90
pytweening==1.2.0
pytz==2024.1
pywhatkit==5.4
PyYAML==6.0.1
pyzmq==26.2.0
Quart==0.11.5
Quart-CORS==0.3.0
referencing==0.35.1
regex==2024.11.6
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.22.3
safetensors==0.5.2
Send2Trash==1.8.3
simpleaudio==1.0.4
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.6
soupsieve==2.5
SpeechRecognition==3.10.0
stack-data==0.6.3
swagger-ui-py==0.2.1
sympy==1.13.1
taskgroup==0.0.0a4
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
terminado==0.18.1
tf_keras==2.18.0
tinycss2==1.4.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.0.1
torch==2.5.1
tornado==6.4.2
tqdm==4.48.2
traitlets==5.14.3
transformers==4.48.0
triton==3.1.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.11.0
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
whichcraft==0.6.1
wikipedia==1.4.0
wolframalpha==5.0.0
wrapt==1.16.0
wsproto==1.2.0
xmltodict==0.13.0
zope.event==5.0
zope.interface==6.4.post2
